backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
qdrant-client==1.9.1
sentence-transformers==3.0.1
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
firecrawl-py==1.5.0
sse-starlette==2.1.0
numpy==1.26.4
# llama.cpp Python bindings — handles GGUF models including BitNet quantized weights.
# CPU: pip install llama-cpp-python
# CUDA GPU: CMAKE_ARGS="-DLLAMA_CUDA=on" pip install llama-cpp-python --force-reinstall
# Metal (Apple): CMAKE_ARGS="-DLLAMA_METAL=on" pip install llama-cpp-python --force-reinstall
llama-cpp-python
